requirements.txt
fastapi
uvicorn
python-multipart
streamlit
tensorflow==2.15.0
keras==2.15.0
pillow
numpy
matplotlib
